// src/models/userModels/user.types.ts
export interface IUser {
    firstName: string
    lastName: string
    email: string
    street: string
    number: number
    city: string
    district: string
    state: string
    birthDate: string
    gender: string
    createdAt: string
    updatedAt: string
    password: string
    phone: string
    facebook_id: number
}